+++ src/Farmer/Builders/Builders.KeyVault.fs
@@ -7,7 +7,6 @@ open Farmer.Arm.KeyVault
 open System
 open Vaults
 
-type NonEmptyList<'T> = 'T * 'T List
 type AccessPolicyConfig =
     { ObjectId : ArmExpression
       ApplicationId : Guid option
@@ -118,17 +117,17 @@ type KeyVaultConfig =
                     let policies =
                         match this.Policies with
                         | Unspecified policies -> policies
-                        | Recover(policy, secondaryPolicies) -> policy :: secondaryPolicies
+                        | Recover list -> list.Value
                         | Default policies -> policies
                     [ for policy in policies do
-                       {| ObjectId = policy.ObjectId
-                          ApplicationId = policy.ApplicationId
-                          Permissions =
-                           {| Certificates = policy.Permissions.Certificates
-                              Storage = policy.Permissions.Storage
-                              Keys = policy.Permissions.Keys
-                              Secrets = policy.Permissions.Secrets |}
-                       |}
+                        {| ObjectId = policy.ObjectId
+                           ApplicationId = policy.ApplicationId
+                           Permissions =
+                            {| Certificates = policy.Permissions.Certificates
+                               Storage = policy.Permissions.Storage
+                               Keys = policy.Permissions.Keys
+                               Secrets = policy.Permissions.Secrets |}
+                        |}
                     ]
                   Uri = this.Uri
                   DefaultAction = this.NetworkAcl.DefaultAction
@@ -247,8 +246,8 @@ type KeyVaultBuilder() =
             match state.CreateMode, state.Policies with
             | None, policies -> Unspecified policies
             | Some SimpleCreateMode.Default, policies -> Default policies
-            | Some SimpleCreateMode.Recover, primary :: secondary -> Recover(primary, secondary)
             | Some SimpleCreateMode.Recover, [] -> failwith "Setting the creation mode to Recover requires at least one access policy. Use the accessPolicy builder to create a policy, and add it to the vault configuration using add_access_policy."
+            | Some SimpleCreateMode.Recover, policies -> Recover (NonEmptyList.create policies)
           Secrets = state.Secrets
           Uri = state.Uri
           Tags = state.Tags  }